Major refactoring
This refactoring decouples the SQL and MyBatis3 implementations from the
base QBE framework.  This should make it possible to write extensions of
the basic framework for other needs.

In addition, this made it possible to write a pure JDBC implementation
that produces where clauses ready for input to plain prepared
statements.