Improve the insert selective API
Change the insert selective API so that all you have to do is specify
the property name.  This requires us to check the value of a property
when composing the insert statement.  To do this, we include a portion
of the MyBatis code for reflecting and reading properties.  This code is
included with the shade plugin to create an uber Jar, and the classes
are relocated so they don't conflict with actual MyBatis.