Yet another strategy for dealing with table aliases.
Create a specialized object AliasMap and pass that into the
SelectListItem functions.  This removes the requirement to expose the
table in SelectListItem which was Optional<SqlTable> only because of the
count all aggregate.  Using an Optional<SqlTable> for only one special
case caused some tortured processing further down the line.