Add Alternative Join Construction Methods
Added a few methods to QueryExpressionDSL to allow building joins with a
single method. This allows us to remove the weird CompletableQuery
interface and makes the forthcoming Kotlin DSL easier to implement.

With this addition, joins can be coded in a single method, rather than
the current method chaining technique:

```java
.join(SomeTable, on(a, equalTo(b)))
```

This does not alter or remove the existing method chaining technique.