Fix for Kotlin
This change removes the old extension methods in favor of builder classes. This is important because there was a possibility of incorrect code generating if the wrong imports were used. The old method had such highly overloaded methods names that the wrong method could have easily been specified which would cause unpredictable results.