Update insert bindings for record based inserts
Record based inserts (insert, batch insert, and
multirow insert) have a different form of
binding - rather than binding to a value
in a Map, these bindings bind to a
property in a row class. This commit makes
the bindings consistent for those statements.

Spring support is also updated to support
the new binding form.