Refactoring Slice - Empty Conditions
This change separates the concept of an empty condition from that of a
renderable condition. This will enable a future change where conditions
could decide to allow rendering even if they are considered empty (such
as rendering empty lists). This change should be transparent to users
unless they have implemented custom conditions.